+++ CHANGELOG.md
@@ -79,6 +79,7 @@
 - [10286](https://github.com/vegaprotocol/vega/issues/10286) - If probability of trading is less than or equal to the minimum, assign it weight of zero for liquidity score calculation and change the validation of the tau scaling network parameter. 
 - [10376](https://github.com/vegaprotocol/vega/issues/10376) - Add spam protection for update profile.
 - [10464](https://github.com/vegaprotocol/vega/issues/10464) - Add total of members in teams API.
+- [10464](https://github.com/vegaprotocol/vega/issues/10464) - Add total of members in referral sets API.
 
 ### 🐛 Fixes